# app/api/API_spec.yaml
openapi: 3.0.3
info:
  title: Digitization Toolkit Backend API
  description: |
    API for the Digitization Toolkit - an open-source, modular digitization platform
    for capturing, managing, and organizing archival materials using Raspberry Pi and 64MP ArduCam cameras.
    
    The API provides endpoints for:
    - User authentication and account management
    - Project and collection management
    - Record (document/archival object) metadata management
    - Image capture control and processing
    - Camera device discovery and calibration
    - System monitoring
  version: 1.0.0
  contact:
    name: UCSB AMP Lab
    url: https://github.com/UCSB-AMPLab/digitization-toolkit-backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://localhost:3000
    description: Production server (frontend proxy)
...
paths:
...
  /collections:
    post:
      summary: Create a new collection
      operationId: create_collection
      tags:
        - collections
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionRead'
        '400':
          description: Invalid parent specification
        '404':
          description: Parent project or collection not found

    get:
      summary: List collections
      operationId: list_collections
      tags:
        - collections
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
          description: Filter by project
        - name: parent_collection_id
          in: query
          schema:
            type: integer
          description: Filter by parent collection
        - name: skip
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectionRead'
  /collections/{collection_id}:
    get:
      summary: Get a specific collection
      operationId: get_collection
      tags:
        - collections
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionRead'
        '404':
          description: Collection not found

    patch:
      summary: Update a collection
      operationId: update_collection
      tags:
        - collections
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Collection updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionRead'
        '400':
          description: Invalid update (e.g., circular hierarchy)
        '404':
          description: Collection not found

    delete:
      summary: Delete a collection
      operationId: delete_collection
      tags:
        - collections
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Collection deleted successfully
        '404':
          description: Collection not found
  /collections/{collection_id}/hierarchy:
    get:
      summary: Get collection with full hierarchy tree
      operationId: get_collection_hierarchy
      tags:
        - collections
      security:
        - BearerAuth: []
      parameters:
        - name: collection_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Collection with nested children
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionWithChildren'
        '404':
          description: Collection not found

  # ============================================================================
  # Record Endpoints
  # ============================================================================
...
components:
  schemas:
...
    CollectionCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        collection_type:
          type: string
          example: series
        archival_metadata:
          type: object
          nullable: true
        project_id:
          type: integer
          nullable: true
        parent_collection_id:
          type: integer
          nullable: true
        created_by:
          type: string
          nullable: true
    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        collection_type:
          type: string
        archival_metadata:
          type: object
        parent_collection_id:
          type: integer
          nullable: true
    CollectionRead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        collection_type:
          type: string
        archival_metadata:
          type: object
        project_id:
          type: integer
          nullable: true
        parent_collection_id:
          type: integer
          nullable: true
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
    CollectionWithChildren:
      allOf:
        - $ref: '#/components/schemas/CollectionRead'
        - type: object
          properties:
            child_collections:
              type: array
              items:
                $ref: '#/components/schemas/CollectionWithChildren'
            record_count:
              type: integer

    # Record Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
...